Fixing warnings for 64bit compilation

+++ axiom/include/axiom_mime_part.h
@@ -71,7 +71,7 @@ extern "C"
         /* Size of the part. In the case of buffer this is 
          * the buffer size and in the case of file this is 
            the file size */
-        int part_size;    
+        size_t part_size;
 
         /* This is one from the above defined enum */
         axiom_mime_part_type_t type;
